+++ src/hybran/main.py
@@ -17,25 +17,86 @@ from . import \
     converter, \
     config, \
     designator, \
+    standardize, \
     __version__
+
+from .argparse import DefaultSubcommandArgumentParser
 
 
 def cmds():
     """
-    argparse parse input provided by the user
-
-    :return: argparse.parse_args() object
+    argparse parse input provided by the user and call appropriate command
     """
-    parser = argparse.ArgumentParser(description='Hybran: hybrid reference-based and ab initio prokaryotic genomic annotation. '
-                                                 'Mixing different species within a single annotation run is NOT recommended.',
-                                     epilog=
-                                     """
-                                     Elghraoui, A.; Gunasekaran, D.; Ramirez-Busby, S. M.; Bishop, E.; Valafar, F.
-                                     Hybran: Hybrid Reference Transfer and Ab Initio Prokaryotic Genome Annotation.
-                                     bioRxiv November 10, 2022, p 2022.11.09.515824.
-                                     <https://doi.org/10.1101/2022.11.09.515824>
-                                     """,
-                                     formatter_class=argparse.ArgumentDefaultsHelpFormatter)
+    citation = (
+        "Elghraoui, A.; Gunasekaran, D.; Ramirez-Busby, S. M.; Bishop, E.; Valafar, F. "
+        "Hybran: Hybrid Reference Transfer and Ab Initio Prokaryotic Genome Annotation. "
+        "bioRxiv November 10, 2022, p 2022.11.09.515824. "
+        "<https://doi.org/10.1101/2022.11.09.515824>"
+    )
+
+    head_parser = DefaultSubcommandArgumentParser(
+        description='Hybran: hybrid reference-based and ab initio prokaryotic genomic annotation.',
+        epilog=citation,
+        formatter_class=argparse.ArgumentDefaultsHelpFormatter
+    )
+
+    subparsers = head_parser.add_subparsers(
+        dest='subparser_name',
+    )
+    # We want the main annotation command to run as default, but also provide alternative commands.
+    #
+    # Standard argparse seems to handle this, with this `required` option to `add_subparsers`,
+    # but it's not actually what we want.
+    #  > required - Whether or not a subcommand must be provided, by default False (added in 3.7)
+    # subparsers = parser.add_subparsers(
+    #     title="Alternative subcommands",
+    #     required=False,
+    # )
+
+    # main command "annotate"
+    parser = subparsers.add_parser(
+        'annotate',
+        help='Run the Hybran annotation pipeline. (default)',
+    )
+    head_parser.set_default_subparser('annotate')
+
+
+    # alternative commands
+    stdize = subparsers.add_parser(
+        'standardize',
+        help='Apply standard naming conventions to Hybran output.'
+    )
+    stdize.set_defaults(func=standardize.main)
+
+
+    #
+    # hybran standardize
+    #
+    stdize.add_argument(
+        'annotations',
+        help="Directory, space-separated list of GBKs, or a FOFN containing all annotated genomes.",
+        nargs='+'
+    )
+    stdize.add_argument(
+        '-p', '--orf-prefix',
+        type=str,
+        help='prefix for generic gene names (*not* locus tags)',
+        default='ORF',
+    )
+    stdize.add_argument(
+        '-o', '--output',
+        help='Directory to output all new annotation files.',
+        default='.',
+    )
+    stdize.add_argument(
+        '-d', '--duplicates-file',
+        help="duplicates.tsv file produced during Hybran's reference deduplication step.",
+        required=True,
+    )
+
+    #
+    # hybran annotate
+    #
     required = parser.add_argument_group('Required')
     optional = parser.add_argument_group('Optional')
     ratt_params = parser.add_argument_group('RATT Options.\n(See http://ratt.sourceforge.net/documentation.html and\n https://github.com/ThomasDOtto/ratt/blob/master/ratt.1.md for more details)')
@@ -56,9 +117,6 @@ def cmds():
                                                            'download_eggnog_data.py -y bactNOG. Full path only',
                           dest='database_dir',
                           required=False)
-    optional.add_argument('--dedupe-references',
-                          action='store_true',
-                          help='Identify duplicate genes in the reference annotations and assign one name to all copies.')
     optional.add_argument('-t', '--first-reference', required=False, dest='first_gbk',
                           help='Reference to use as the reference database for Prokka. Must exist in --references dir.'
                                ' Default is the first reference annotation (Genbank) in -r/--references.')
@@ -155,7 +213,12 @@ def cmds():
                                default=1e-9,
                                help='Similarity e-value cut-off')
 
-    arguments = parser.parse_args()
+    arguments = head_parser.parse_args()
+
+    if arguments.subparser_name != 'annotate':
+        arguments.func(arguments)
+        return
+
     # Turn Prokka arguments back into an option string that we can pass directly to Prokka,
     # rather than having to repopulate --kingdom args.kingdom --gram args.gram ...
     #
@@ -166,10 +229,10 @@ def cmds():
          for g in parser._action_groups[-1]._group_actions if vars(arguments)[g.dest]]
     ).replace("True","")
 
-    return arguments, prokka_passthrough
+    main(arguments, prokka_passthrough)
 
 
-def main():
+def main(args, prokka_args):
     """
     Hybran: a pipeline to annotate Mycobacterium
     tuberculosis de novo assembled genomes. Annotation of other species
@@ -177,8 +240,6 @@ def main():
 
     :return: None
     """
-    global args
-    args, prokka_args = cmds()
     # Obtaining the absolute path to all scripts
     script_dir = os.path.abspath(os.path.dirname(__file__))
 
@@ -237,13 +298,12 @@ def main():
     os.chdir(args.output)
 
     # Setting up references for RATT, as well as versions in GFF format used later
-    if args.dedupe_references:
-        if not os.path.isdir('deduped-refs'):
-            try:
-                os.mkdir('deduped-refs')
-            except:
-                sys.exit("Could not create directory: deduped-refs ")
-        args.references = refManager.dedupe(args.references, outdir='deduped-refs', tmpdir=hybran_tmp_dir)
+    if not os.path.isdir('deduped-refs'):
+        try:
+            os.mkdir('deduped-refs')
+        except:
+            sys.exit("Could not create directory: deduped-refs ")
+    args.references = refManager.dedupe(args.references, outdir='deduped-refs', tmpdir=hybran_tmp_dir)
     refdir, embl_dir, embls = fileManager.prepare_references(args.references)
     intermediate_dirs = ['clustering/', 'eggnog-mapper-annotations/', 'prodigal-test/', refdir] + \
                         [d for d in glob.glob('emappertmp*/')]
@@ -363,4 +423,4 @@ def main():
 
 
 if __name__ == '__main__':
-    main()
+    cmds()